+++ HISTORY.rst
@@ -6,6 +6,9 @@ Pending
 
 .. Insert new release notes below this line
 
+* Support the value `file://` for origins which is accidentally sent by some
+  versions of Chrome on Android.
+
 3.1.0 (2019-08-13)
 ------------------